dom/tests/mochitest/bugs/test_bug260264.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=260264
-->
<head>
  <title>Test for Bug 260264</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="application/javascript" src="utils_bug260264.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=260264">Mozilla Bug 260264</a>
<p id="display">
  <a id="link" href="javascript:(function(){})()">link</a>
</p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 260264 **/

SimpleTest.requestFlakyTimeout("untriaged");

var a = $("link"),
    checkOpened = function() { ok(window.open("http://example.com"), "not properly opened") },
    checkBlocked = function() { ok(!window.open("http://example.com"), "not properly blocked") };

const popupSpamMaximum = 3;
const eventInAllowedList = "click";
const eventNotInAllowedList = "mouseup";

async function testPopupSpamMaximum() {
  // Opening a popup from an event in the allowed event list doesn’t affect the
  // popup spam count.
  for (var count = 0, n = 0; n < popupSpamMaximum; n++) {
    sendWithUserActivation(a, eventInAllowedList, checkOpened);
  }
  sendWithUserActivation(a, eventInAllowedList, checkOpened);

  for (var count = 0, n = 0; n < popupSpamMaximum; n++) {
    sendWithUserActivation(a, eventNotInAllowedList, checkOpened);
  }
  sendWithUserActivation(a, eventNotInAllowedList, checkBlocked);

  // Close one spam popup.
  await window.open.close(1);
  sendWithUserActivation(a, eventNotInAllowedList, checkOpened);
  sendWithUserActivation(a, eventNotInAllowedList, checkBlocked);
  // Still allow opening a popup from an event in the allowed event list even if
  // the popup spam count has reached its maximum.
  sendWithUserActivation(a, eventInAllowedList, checkOpened);

  await window.open.close();
}

add_setup(async function init() {
  await SpecialPowers.pushPrefEnv({"set": [
    // Enable popup blocker
    ["dom.disable_open_during_load", true],
    ["dom.popup_allowed_events", eventInAllowedList],
    ["dom.popup_maximum", popupSpamMaximum],
  ]});
});

add_task(async function test_unknown_permission() {
  await SpecialPowers.pushPermissions([{'type': 'popup', 'remove': true, 'context': document}]);
  ok(await SpecialPowers.testPermission('popup', UNKNOWN_ACTION, document), 'popup value should have UNKNOWN permission');

  await testPopupSpamMaximum();

  await SpecialPowers.popPermissions();
})

add_task(async function run_allow_permission() {
  await SpecialPowers.pushPermissions([{'type': 'popup', 'allow': ALLOW_ACTION, 'context': document}]);
  ok(await SpecialPowers.testPermission('popup', ALLOW_ACTION, document), 'popup value should have ALLOW permission');

  await testPopupSpamMaximum();

  await SpecialPowers.popPermissions();
});

add_task(async function run_deny_permission() {
  await SpecialPowers.pushPermissions([{'type': 'popup', 'allow': DENY_ACTION, 'context': document}]);
  ok(await SpecialPowers.testPermission('popup', DENY_ACTION, document), 'popup value should have ALLOW permission');

  await testPopupSpamMaximum();

  await SpecialPowers.popPermissions();
});

</script>
</pre>
</body>
</html>
